requirements.txt
requests>=2.28.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentencepiece>=0.1.99
